refactor(tests): Moved test messages structs to tests helper directory
